backend/requirements.txt
fastapi==0.136.1
uvicorn[standard]==0.46.0
gunicorn==22.0.0
pydantic==2.13.3
pydantic-settings==2.10.1
sqlalchemy[asyncio]==2.0.49
asyncpg==0.31.0
alembic==1.18.4
celery[redis]==5.6.3
redis==7.4.0
cryptography==47.0.0
python-dotenv==1.2.2
httpx==0.28.1
pytest==8.3.5
pytest-asyncio==0.25.3
pytest-cov==6.1.0
pyjwt[crypto]>=2.10.1
pwdlib[argon2]==0.2.1
python-multipart==0.0.20
